[FEATURE] make it easier to configure templates.
Extensions providing base templates can now use the `template`
function to add extra templates to the config. Templates
must be added in the `templates` section of the guides
configuration.
