+++ guides.xml
@@ -11,7 +11,6 @@
         log-path="php://stder"
 >
     <project title="phpDocumentor Guides"/>
-
     <extension class="phpDocumentor\Guides\Bootstrap"/>
     <extension class="phpDocumentor\Guides\Code"/>
 </guides>
