+++ phpstan-baseline.neon
@@ -1,11 +1,6 @@
 parameters:
 	ignoreErrors:
 		-
-			message: "#^Parameter \\#1 \\$value of function count expects array\\|Countable, iterable\\<Symfony\\\\Component\\\\Console\\\\Command\\\\Command\\> given\\.$#"
-			count: 1
-			path: packages/guides-cli/src/Application.php
-
-		-
 			message: "#^Call to function is_countable\\(\\) with array will always evaluate to true\\.$#"
 			count: 2
 			path: packages/guides-cli/src/Command/Run.php
